Fixed some minor issues from the last merge.
Added the SprintContext class, which was found online, it allows the access of a productFileDAO instance in the buyerFileDAO instance so the total cost of a users cart could be calculated. Additionally, two buyer tests were just commented out for the time being because buyer will be shortly changed to buyerInfo, and all its tests with it.